hello_world/requirements.txt
mesop
gunicorn
google-generativeai
google-cloud-bigquery
faker
sqlparse